backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.1
reportlab==4.2.2
google-generativeai==0.8.2
SQLAlchemy==2.0.36
requests==2.31.0
python-multipart==0.0.20
